+++ src/main/java/tonius/simplyjetpacks/integration/RAItems.java
@@ -1,6 +1,7 @@
 package tonius.simplyjetpacks.integration;
 
 import net.minecraft.item.ItemStack;
+import net.minecraftforge.oredict.OreDictionary;
 import tonius.simplyjetpacks.SimplyJetpacks;
 import cpw.mods.fml.common.registry.GameRegistry;
 
@@ -13,7 +14,7 @@ public abstract class RAItems {
         SimplyJetpacks.logger.info("Stealing Redstone Arsenal's items");
         
         plateFlux = GameRegistry.findItemStack("RedstoneArsenal", "plateFlux", 1);
-        armorFluxPlate = GameRegistry.findItemStack("RedstoneArsenal", "armorFluxPlate", 1);
+        armorFluxPlate = new ItemStack(GameRegistry.findItemStack("RedstoneArsenal", "armorFluxPlate", 1).getItem(), 1, OreDictionary.WILDCARD_VALUE);
     }
     
 }