+++ VersionHistory.md
@@ -2,8 +2,9 @@
 ----------------------------------
 - Added jetpack sounds. Many thanks to nabijaczleweli for help with this one. The volume for these sounds can be set with the volume slider for the 'Players' sound category. They can also be disabled as a whole in the config (clientside).
 - All jetpack/flux pack types can now actually be tuned individually (instead of being bound to a list of tiers).
+- The Flux-Infused JetPlate is now craftable even if Redstone Arsenal is not installed.
 - Localized keybind names. The mod's keybinds will likely be reset when updating to this version.
-- Removed 'flat' jetpack armor models. This is much cleaner internally.
+- Removed 'flat' jetpack armor models.
 - Texture pack authors, you'll have to rename some texture files: tuberous/creative/icon jetpacks, creative flux packs, and 3d jetpack armor models.
 - Jetpack particles now spread out a tiny little bit, instead of going straight down.
 - Creative Jetpacks now have rainbow smoke particles by default.